requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
future==1.0.0
pillow==10.4.0
serializers==0.2.4
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
